Ajustando las ventas, dialogo de datos de facturacion listo, procesando la seleccion de libros.

+++ src/java/controlador/jsf/DetVentaController.java
@@ -1,6 +1,5 @@
 package controlador.jsf;
 
-
 import controlador.jsf.util.JsfUtil;
 import controlador.jsf.util.JsfUtil.PersistAction;
 import java.io.Serializable;
@@ -18,7 +17,6 @@ import javax.faces.convert.FacesConverter;
 import javax.inject.Named;
 import modelo.DetVenta;
 
-
 @Named("detVentaController")
 @SessionScoped
 public class DetVentaController implements Serializable {
@@ -31,39 +29,32 @@ public class DetVentaController implements Serializable {
     public DetVentaController() {
     }
 
-
     public DetVenta getSelected() {
         return selected;
     }
-
 
     public void setSelected(DetVenta selected) {
         this.selected = selected;
     }
 
-
     protected void setEmbeddableKeys() {
-        selected.getDetVentaPK().setIdLibro(selected.getLibro().getIdLibro());
+        selected.getDetVentaPK().setIdLibro(selected.getLibro().getIdLibro().toBigInteger());
         selected.getDetVentaPK().setIdVenta(selected.getVenta().getIdVenta());
     }
-
 
     protected void initializeEmbeddableKey() {
         selected.setDetVentaPK(new modelo.DetVentaPK());
     }
 
-
     private DetVentaFacade getFacade() {
         return ejbFacade;
     }
-
 
     public DetVenta prepareCreate() {
         selected = new DetVenta();
         initializeEmbeddableKey();
         return selected;
     }
-
 
     public void create() {
         persist(PersistAction.CREATE, ResourceBundle.getBundle("/Bundle").getString("DetVentaCreated"));
@@ -72,11 +63,9 @@ public class DetVentaController implements Serializable {
         }
     }
 
-
     public void update() {
         persist(PersistAction.UPDATE, ResourceBundle.getBundle("/Bundle").getString("DetVentaUpdated"));
     }
-
 
     public void destroy() {
         persist(PersistAction.DELETE, ResourceBundle.getBundle("/Bundle").getString("DetVentaDeleted"));
@@ -86,14 +75,12 @@ public class DetVentaController implements Serializable {
         }
     }
 
-
     public List<DetVenta> getItems() {
         if (items == null) {
             items = getFacade().findAll();
         }
         return items;
     }
-
 
     private void persist(PersistAction persistAction, String successMessage) {
         if (selected != null) {
@@ -123,21 +110,17 @@ public class DetVentaController implements Serializable {
         }
     }
 
-
     public DetVenta getDetVenta(modelo.DetVentaPK id) {
         return getFacade().find(id);
     }
-
 
     public List<DetVenta> getItemsAvailableSelectMany() {
         return getFacade().findAll();
     }
 
-
     public List<DetVenta> getItemsAvailableSelectOne() {
         return getFacade().findAll();
     }
-
 
     @FacesConverter(forClass = DetVenta.class)
     public static class DetVentaControllerConverter implements Converter {
@@ -155,7 +138,6 @@ public class DetVentaController implements Serializable {
             return controller.getDetVenta(getKey(value));
         }
 
-
         modelo.DetVentaPK getKey(String value) {
             modelo.DetVentaPK key;
             String values[] = value.split(SEPARATOR_ESCAPED);
@@ -165,7 +147,6 @@ public class DetVentaController implements Serializable {
             return key;
         }
 
-
         String getStringKey(modelo.DetVentaPK value) {
             StringBuilder sb = new StringBuilder();
             sb.append(value.getIdVenta());
@@ -173,7 +154,6 @@ public class DetVentaController implements Serializable {
             sb.append(value.getIdLibro());
             return sb.toString();
         }
-
 
         @Override
         public String getAsString(FacesContext facesContext, UIComponent component, Object object) {
@@ -188,7 +168,6 @@ public class DetVentaController implements Serializable {
                 return null;
             }
         }
-
 
     }
